+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+
+* Method `firstToLower()` - converts first character in the string to lower case.
+
 ## [0.1.1] - 2021-04-07
 
 * Method `chunks()` - splits string by chunk with defined size.